Refactor json media type method and move it to ContentType
