+++ logbook-json/src/main/java/org/zalando/logbook/json/CompactingJsonBodyFilter.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apiguardian.api.API;
 import org.zalando.logbook.BodyFilter;
-import org.zalando.logbook.internal.JsonMediaType;
+import org.zalando.logbook.ContentType;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
@@ -27,7 +27,7 @@ public final class CompactingJsonBodyFilter implements BodyFilter {
 
     @Override
     public String filter(@Nullable final String contentType, final String body) {
-        if (!JsonMediaType.JSON.test(contentType)) {
+        if (!ContentType.isJsonMediaType(contentType)) {
             return body;
         }
